TIP-874: move category updater
